Add GaugedBalancingPolicy >: MeteredBalancingPolicy
